// src/codegen/models/ListCommentsResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { CommentResponse } from './CommentResponse';

export type ListCommentsResponse = {
  /**
   * Indicates whether the operation was successful or not
   */
  success: boolean;
  /**
   * List of comments
   */
  data: Array<CommentResponse>;
  /**
   * If the operation was not successful, holds the deatils of errors
   */
  errors: Array<string>;
};