// src/codegen/models/UpdateMediaRequestBody.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UpdateMediaRequestBody = {
  /**
   * Name of the media file
   */
  name?: string;
  /**
   * Project ID to associate the media with
   */
  projectId?: string;
  /**
   * Folder ID to move the media to
   */
  folderId?: string;
};